Issue #2 More refactoring
