+++ package.json
@@ -17,8 +17,7 @@
   "dependencies": {
     "axios": "^1.5.0",
     "gl-matrix": "^3.4.3",
-    "satellite.js": "^5.0.0",
-    "spin.js": "^4.1.1"
+    "satellite.js": "^5.0.0"
   },
   "devDependencies": {
     "eslint": "^8.49.0",
