+++ public/images/icons/attribution.md
@@ -1,0 +1,1 @@
+Icons are from https://fonts.google.com/icons
