+++ src/constants.js
@@ -1,6 +1,16 @@
 /* eslint-disable import/prefer-default-export */
 const R2D = 80 / Math.PI;
 
+const Events = {
+  satMovementChange: 'satMovementChange',
+  selectedSatChange: 'selectedSatChange',
+  satHover: 'sathover',
+  satDataLoaded: 'satdataloaded',
+  closeWindow: 'closeWindow',
+  cruncherReady: 'cruncherReady'
+};
+
 export {
-  R2D
+  R2D,
+  Events
 };
